requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
alabaster==0.7.12
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==1.10.0
ansi2html==1.8.0
appdirs==1.4.4
argh==0.26.2
astor==0.8.1
asttokens==2.4.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
backports.weakref==1.0.post1
bar-chart-race==0.1.0
base58==2.1.0
bkcharts==0.2
black==19.10b0
blessings==1.7
blinker==1.4
blis==0.7.4
boto==2.49.0
boto3==1.4.8
botocore==1.8.50
Bottleneck==1.3.2
branca==0.4.2
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.10.8
chart-studio==1.1.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.1
clikit==0.6.2
clyent==1.2.2
colorlover==0.3.0
conda==4.10.1
conda-build==3.21.4
conda-verify==3.4.2
confi==0.0.4.1
confuse==1.4.0
crashtest==0.3.1
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
cytoolz==0.11.0
dash==2.14.0
dash-bootstrap-components==0.12.2
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dask==2021.10.0
deprecation==2.1.0
descartes==1.1.0
dfsql==0.3.1
dill==0.3.2
distlib==0.3.1
dtale==1.49.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==2.0.0
fastcache==1.1.0
ffmpeg==1.4
Fiona==1.8.18
Flask-Compress==1.10.1
Flask-Cors==3.0.8
flask-ngrok==0.0.25
Flask-SQLAlchemy==2.4.3
flatbuffers==1.12
folium==0.12.1
future==0.18.2
gast==0.4.0
gensim==4.0.1
geographiclib==2.0
geopandas==0.8.2
geopy==2.4.0
gitdb==4.0.7
GitPython==3.1.18
gmpy2==2.0.8
google-api-core==1.30.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.34.1
gunicorn==20.0.4
heroku==0.1.4
hiredis==2.0.0
htmlmin==0.1.12
httpstan==4.5.0
icecream==2.1.3
ImageHash==4.2.0
importlib-resources==5.1.4
jmespath==0.10.0
jupyter-kite==2.0.2
jupyter-packaging==0.9.2
jupyter-server==1.6.4
kaleido==0.2.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
llvmlite==0.36.0
locket==0.2.1
LunarCalendar==0.0.9
lux==0.8.2
lz4==3.1.3
Markdown==3.3.4
marshmallow==3.12.1
mccabe==0.6.1
mindsdb-sql==0.0.12
missingno==0.4.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
modin==0.10.0
modin-spreadsheet==0.1.2
mpmath==1.2.1
multidict==5.1.0
multimethod==1.4
multipledispatch==0.6.0
multiprocess==0.70.10
munch==2.5.0
murmurhash==1.0.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
navigator-updater==0.2.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.5.3.56
opt-einsum==3.3.0
p5py==1.0.0
pandas==1.2.4
pandas-datareader==0.9.0
pandas-profiling==2.11.0
pandas-visual-analysis==0.0.4
parso==0.7.0
pastel==0.2.1
pathos==0.2.6
pathspec==0.7.0
pathy==0.5.2
patsy==0.5.1
pep517==0.10.0
pep8==1.7.1
phik==0.11.0
plotly==5.17.0
plumbum==1.7.0
ply==3.11
pox==0.2.8
ppft==1.6.6.2
ppscore==1.2.0
preshed==3.0.5
protobuf==3.15.5
psycopg2==2.7.7
pulsar==2.0.2
pulsar-odm==0.7.0
py-spy==0.3.7
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycrypto==2.6.1
pycurl==7.44.1
pydantic==1.7.4
pydeck==0.6.2
pydot==1.3.0
pyforest==1.0.3
PyJWT==2.1.0
pylev==1.4.0
pymongo==3.10.1
PyMySQL==1.0.2
pyodbc===4.0.0-unsupported
pyproj==3.0.0.post1
pysimdjson==3.2.0
pystan==3.2.0
pytest==6.2.4
python-slugify==4.0.1
pytz==2021.3
pywebio==1.3.3
PyYAML==6.0
ray==1.4.0
redis==3.5.3
requests-oauthlib==1.3.0
retrying==1.3.3
rpyc==4.1.5
rsa==4.7.2
s3transfer==0.1.13
scikit-image==0.16.2
scikit-plot==0.3.7
Shapely==1.7.1
simplegeneric==0.8.1
sklearn==0.0
sly==0.4
smart-open==3.0.0
smmap==4.0.0
spacy==3.0.6
spacy-legacy==3.0.6
squarify==0.4.3
srsly==2.4.1
streamlit==0.83.0
strsimpy==0.2.0
sweetviz==2.1.2
tangled-up-in-unicode==0.0.6
tenacity==8.2.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
text-unidecode==1.3
texthero==1.0.9
thinc==8.0.6
tifffile==2020.10.1
tomlkit==0.7.0
tweepy==3.10.0
TwitterSearch==1.0.2
typer==0.3.2
typing_extensions==4.7.1
tzlocal==2.1
ua-parser==0.10.0
unicodecsv==0.14.1
Unidecode==1.2.0
user-agents==2.2.0
validators==0.18.2
virtualenv==20.4.3
visions==0.6.0
wasabi==0.8.2
webargs==7.0.1
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
xgboost==1.4.2
xlwt==1.3.0
yarl==1.6.3
zict==2.0.0
zope.event==4.5.0
